weja tworzy3

+++ operator1.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
 int main() {
-    float a = 4;
+    int a = 4;
     int b = 7;
 
-    printf("%f\n", a++ + b);
-    printf("%f\n", ++a + b);
+    printf("%i\n", a++ + b);
+    printf("%i\n", ++a + b);
 
-    printf("%f\n", --a + b);
-    printf("%f\n", a-- + b);
+    printf("%i\n", --a + b);
+    printf("%i\n", a-- + b);
 
     return 0;
 }
